SAR-3320: Optional postcode for overseas support
update Storevatnumber service, known facts matching service
and eligibility service to support the optional postcode.
This change is needed since overseas companies are not
required to insert one as a known facts, since in VMF
their postcode will be a placeholder, which means that
anything other than the placeholder (which the user isn't
aware of) would cause the known facts service to mismatch.

Note: Some files based on the request id have been removed.